explore opps modification efficiency calculations
